requirements.txt
asgiref==3.11.0
distlib==0.4.0
Django==5.2.9
django-browser-reload==1.21.0
django-tailwind==4.4.2
filelock==3.20.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
pytailwindcss==0.3.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
virtualenv==20.35.4
